// ui/src/lib/booking-hotel-api/model/user-model.ts
/**
 * main-app API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {UserRole} from './user-role';


export interface UserModel {
    id: number;
    firstName: string;
    lastName: string;
    email: string;
    role: UserRole;
    createdAt: string;
    updatedAt: string;
    removed: boolean;
}
export namespace UserModel {
}